scsi: sd: Implement support for NDOB flag in WRITE SAME(16)
The NDOB flag removes the need for a zeroed logical block in the data-out
buffer when using WRITE SAME(16) to zero block ranges.  Implement support
for NDOB in the SCSI disk driver to mirror WRITE ZEROES in NVMe.

The only way to detect whether a device supports NDOB is through REPORT
SUPPORTED OPERATION CODES. Since we can't safely send that command to all
devices we only attempt this if the device implements the Block
Provisioning VPD page and sets the LBPWS flag.

If we issue a WRITE SAME(16) we check whether NDOB is set for the device in
question. If so we do not allocate a zeroed page from the pool and simply
issue the command with a zero-length payload.

Link: https://lore.kernel.org/r/20220302053559.32147-12-martin.petersen@oracle.com
Reviewed-by: Christoph Hellwig <hch@lst.de>
Reviewed-by: Bart Van Assche <bvanassche@acm.org>
Reviewed-by: Johannes Thumshirn <johannes.thumshirn@wdc.com>
Signed-off-by: Martin K. Petersen <martin.petersen@oracle.com>