ipts: Implement new companion interface
This makes it easier to extend the companion driver, since everything is
stored in a single struct.

The ipts_companion_t struct is stored in two places, the IPTS main driver,
and the driver registering it. It is still passed through to the functions
that the companion driver implements to allow for device specific behaviour.

Signed-off-by: Dorian Stoll <dorian.stoll@tmsp.io>