ipts: Let the companion decide about no_feedback
Having a DMI match for some surface devices doesn't really match with
the concept of having a device-specific companion driver, so this commit
moves the matching code to the companion.

ipts-surface is now setup in a way that allows us to add device-specific
data pretty easily. All you need is the interface to IPTS, which can
just be copied from no_feedback.

Signed-off-by: Dorian Stoll <dorian.stoll@tmsp.io>