Fixup: BACKPORT: ipts: add static_assert definition to header files
Fixup the following commits:
    BACKPORT: ipts: Cleanup ipts-sensor-regs.h to pass checkpatch
    BACKPORT: ipts: Clean up ipts-mei-msg.h to pass checkpatch

The static_assert macro is not available on 4.19. So, there are two ways
to fixup those commits:
    1. Backport upstream commit[1] that adds the static_assert
    2. Add the definition to ipts header files

I chose the latter way because backporting the upstream commit causes
conflicts and I have to a) backport other commit(s) or b) resolve the
conflict manually.

Also, the original IPTS patch[2] is doing the same way; they defined
C_ASSERT in those files [3][4].

References:
[1] https://github.com/torvalds/linux/commit/6bab69c65013bed5fce9f101a64a84d0385b3946
[2] https://github.com/ipts-linux-org/ipts-linux-new
[3] https://github.com/ipts-linux-org/ipts-linux-new/blob/master/drivers/misc/ipts/ipts-sensor-regs.h
[4] https://github.com/ipts-linux-org/ipts-linux-new/blob/master/drivers/misc/ipts/ipts-mei-msgs.h