HID: Do not abort on over-sized reports
Commit 8ec321e96e056de84022c032ffea253431a83c3c (HID: Fix
slab-out-of-bounds read in hid_field_extract) causes HID devices to not
be added when they contain a report in the descriptor that is too large.
Instead of refusing to add the device, we just ignore the report. This
is more in spirit of the actual report handling code, which truncates
reports if they are too large.

Based on the patch proposed on bugzilla by 'js'.

Bugzilla: https://bugzilla.kernel.org/show_bug.cgi?id=206259