HID: fix accidentally added line when backporting
When verifying generated patches on v4.19-surface, I noticed a small
difference:

    $ git checkout v4.19.123
    $ git merge v4.19-surface
    $ git rebase v4.19.123
    $ git tag v4.19.123-surface

    $ git checkout v4.19.123
    $ git merge v4.19-surface-devel
    $ git rebase v4.19.123
    $ git tag v4.19.123-surface-devel

    $ git diff v4.19.123-surface-devel v4.19.123-surface
    diff --git a/drivers/hid/hid-ids.h b/drivers/hid/hid-ids.h
    index d13b666ed110..b2fff44c8109 100644
    --- a/drivers/hid/hid-ids.h
    +++ b/drivers/hid/hid-ids.h
    @@ -817,7 +817,6 @@
    #define USB_DEVICE_ID_MS_TOUCH_COVER_2   0x07a7
    #define USB_DEVICE_ID_MS_TYPE_COVER_2    0x07a9
    #define USB_DEVICE_ID_MS_POWER_COVER     0x07da
    -#define USB_DEVICE_ID_MS_XBOX_ONE_S_CONTROLLER 0x02fd
    #define USB_DEVICE_ID_MS_PIXART_MOUSE    0x00cb

    #define USB_VENDOR_ID_MOJO             0x8282

It seems that the line has been accidentally added on commit 529ef08d246f55
("Revert "HID: quirks for Microsoft Surface devices""), which seems to
be a backported commit from v5.3-surface-devel commit 352b9fcff865d3.

That line is not on v4.19-surface (because of how patches are made). So,
let's remove the line from v4.19-surface-devel.

Note: no need to cherry-pick this commit to v4.19-surface.

Fixes: 529ef08d246f55 ("Revert "HID: quirks for Microsoft Surface devices"")
Signed-off-by: Tsuchiya Yuto (kitakar5525) <kitakar@gmail.com>