Revert "PCI: Update platform power state when updating PCI state"
This reverts commit ea52a275867bb5c724dd0d383163b97244e5a579.

It turns out that this causes a problem on 7th generation Surface
devices, i.e. Surface Pro 7, Surface Book 3, and Surface Laptop 3.
Specifically this change causes the Intel LPSS PCI driver to fail when
accessing PCI resources and thus fails to load.

Revert this change to replace it with a fix that doesn't break things.