misc: ipts: Update documentation in protocol.h (again...)
I swear at some point I will be done reformatting this file...

Moved all the stuff (except command / response codes) back to using
enums, which lets be honest, just makes 1000 times more sense here. The
kernel-doc comments I added are also the "canonical" way of doing docs
in the kernel, so using that is probably better.

And if we bump the max line length to 100 instead of 80 it doesn't even
look bad.

Signed-off-by: Dorian Stoll <dorian.stoll@tmsp.io>