media: i2c: ov5693: Add in clk*() functions
We removed the clock functions from our hacky driver, because we didn't
realise one of the GPIO lines from the INT3472 devices is actually a
clock enable pin. This is now handled in the intel_skl_int472 module by
registering a clock for the device, so add the clk*() funcs back in to
the driver.

Signed-off-by: Daniel Scally <djrscally@gmail.com>