media: i2c: Check for software_node secondary on probe
On some platforms the firmware node endpoint could be represented by
a software node assigned as secondary to the device's firmware node.
Check for that situation before deciding to fail probe.

In the event that no endpoint is found, fail with -EPROBE_DEFER rather
than -EINVAL; the software node graph could be created later by the
cio2-bridge, so we should check for that later.

Signed-off-by: Daniel Scally <djrscally@gmail.com>