hack! platform/surface: aggregator_registry: Add support for Surface Pro X
Preliminary support for the Surface Pro X. We still need to check how
hot-plugging of the typecover behaves. Ideally, the respective devices
should be removed when it is detached.

Detaching/attaching seems to be indicated by a TC=0x0E event.

We're also still missing a IID=0x02 device which seems for pen removal.