hack! platform/surface: aggregator: Fix for building as builtin module
There seems to be an issue with module ordering, placing some SAM driver
modules before the SAM core module, causing them to complain that the
respective drivers cannot be registered because the 'surface_aggregator'
bus has not been initialized yet.

So let's use subsys_initcall() for the core module to ensure that it's
always loaded before the others.