Revert "hack: fix EFI boot"
This reverts commit 1dfde018f9460cfa497108ced260268c0fdf5a87.

This is not necessary, we can instead use the 'efi=novamap' command line
parameters on newer kernels to get the device to boot.

Furthermore, this change causes us to not exit the UEFI boot services at
all, which is probably not a good idea. Note, however, that not exiting
boot services also seems to have had some unintentional consequences
that made things easier for us: Specifically, reverting this change
causes some issues with the UARTs, which in turn causes a hang and
following hard-reset of the device. We speculate that this might be due
to some firmware-state that gets erased when exiting boot services.

Unfortunately that means that, until we have diagnosed this, the UART
driver module (qcom_geni_serial) needs to be blacklisted, which in turn
prevents SAM and thus the keyboard/touchpad/battery/... from working.