hack! clk: qcom: gcc: Enable basic clocks for Surface Pro X
WARNING: This is a horrible hack, intended for the Surface Pro X booting
in ACPI mode only. It might break stuff.

Currently booting via ACPI doesn't set enable/set up clocks.

First, allow the GCC driver to probe agains the PEP (QCOM0419) ACPI
device. The MMIO region is unfortunately not declared in ACPI and is
hardcoded into the driver on Windows, so we'll to do the same here.

Second, enable the necessary clocks required for SAM. Note: This only
enables clocks and does not set them up fully. While this now prevents
the device from crashing when accessing QUP registers, communication
still does not work and we receive garbage data.