Revert "x86/i8259: Work around buggy legacy PIC"
This reverts commit 1844b2d2fbd859292140df04fc0a89ff47170bde.

Following the thread [1] on this patch, not having a legacy PIC is
perfectly valid.  Rather than continually prodding the PIC until it
looks like it's functioning, which we can't validate, we can provide a
quirk that tells the IOAPIC how to setup interrupt 7 so pinctrl-amd can
be properly set up.  Normally this override is provided by the ACPI
table, however it seems to be missing for irq 7, it is provided on the
Surface Laptop 4 AMD for int 9 for example.

[1]: https://lore.kernel.org/lkml/20210512210459.1983026-1-luzmaximilian@gmail.com/